chore: product 초기 데이터 함수 분리
