+++ src/basic/main.basic.js
@@ -26,6 +26,7 @@ import {
   POINT_RATES_BULK_BONUS,
   PRODUCT_DEFAULT_DISCOUNT_RATES,
 } from './constants';
+import { initProductList } from './data';
 
 // 상품 데이터 및 장바구니 관련 변수
 let productList;
@@ -819,54 +820,4 @@ function getProductDiscountRate(productId) {
     default:
       return 0;
   }
-}
-
-function initProductList() {
-  return [
-    {
-      id: PRODUCT_KEYBOARD,
-      name: '버그 없애는 키보드',
-      val: 10000,
-      originalVal: 10000,
-      availableStock: 50,
-      onSale: false,
-      suggestSale: false,
-    },
-    {
-      id: PRODUCT_MOUSE,
-      name: '생산성 폭발 마우스',
-      val: 20000,
-      originalVal: 20000,
-      availableStock: 30,
-      onSale: false,
-      suggestSale: false,
-    },
-    {
-      id: PRODUCT_MONITOR_ARM,
-      name: '거북목 탈출 모니터암',
-      val: 30000,
-      originalVal: 30000,
-      availableStock: 20,
-      onSale: false,
-      suggestSale: false,
-    },
-    {
-      id: PRODUCT_LAPTOP_POUCH,
-      name: '에러 방지 노트북 파우치',
-      val: 15000,
-      originalVal: 15000,
-      availableStock: 0,
-      onSale: false,
-      suggestSale: false,
-    },
-    {
-      id: PRODUCT_SPEAKER,
-      name: `코딩할 때 듣는 Lo-Fi 스피커`,
-      val: 25000,
-      originalVal: 25000,
-      availableStock: 10,
-      onSale: false,
-      suggestSale: false,
-    },
-  ];
 }
